REFACTORY BACKEND. TODAS ENTIDADES FUNCIONAN CON ID DE TIPO INTEGER, RESOLUCION DE MAPEO DE ENTIDADES, cambios efectuados en DTOS, reestructuracion de service-presenter-repository de cada entidad.

+++ backend/src/main/java/unpsjb/labprog/backend/business/repository/BloqueHorarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import unpsjb.labprog.backend.model.BloqueHorario;
 
-public interface BloqueHorarioRepository extends CrudRepository<BloqueHorario, Long> {
-    List<BloqueHorario> findByAgendaId(Long agendaId);
+public interface BloqueHorarioRepository extends CrudRepository<BloqueHorario, Integer> {
+    List<BloqueHorario> findByAgendaId(Integer agendaId);
 }
