+++ backend/src/main/java/unpsjb/labprog/backend/business/repository/DisponibilidadMedicoRepository.java
@@ -1,12 +1,13 @@
 package unpsjb.labprog.backend.business.repository;
 
+import java.time.LocalTime;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import unpsjb.labprog.backend.model.DisponibilidadMedico;
 import unpsjb.labprog.backend.model.StaffMedico;
 
-import java.time.LocalTime;
-
-public interface DisponibilidadMedicoRepository extends JpaRepository<DisponibilidadMedico, Long> {
+public interface DisponibilidadMedicoRepository extends JpaRepository<DisponibilidadMedico, Integer> {
     boolean existsByStaffMedicoAndDiaSemanaAndHoraInicioAndHoraFin(
         StaffMedico staffMedico, String diaSemana, LocalTime horaInicio, LocalTime horaFin);
 }
